+++ SniffMeet/SniffMeet/Source/Scene/Auth/SetProfileInfo/Interactor/ProfileSetInteractor.swift
@@ -7,34 +7,51 @@
 
 import UIKit
 
-protocol ProfileCreateInteractable: AnyObject {
+protocol ProfileSetInteractable: AnyObject {
     var presenter: (any DogInfoInteractorOutput)? { get set }
     var saveUserInfoUseCase: any SaveUserInfoUseCase { get set }
     var saveProfileImageUseCase: any SaveProfileImageUseCase { get }
     
     func signInWithProfileData(dogInfo: UserInfo, imageData: Data?)
     func convertImageToJPGData(image: UIImage?) -> Data?
+    func isNicknameTaken(_ nickname: String)
 }
 
-final class ProfileCreateInteractor: ProfileCreateInteractable {
+final class ProfileSetInteractor: ProfileSetInteractable {
     weak var presenter: (any DogInfoInteractorOutput)?
     var saveUserInfoUseCase: any SaveUserInfoUseCase
     var saveProfileImageUseCase: any SaveProfileImageUseCase
     var saveUserInfoRemoteUseCase: any CreateAccountUseCase
     var signInUseCase: any SignInUseCase
+    var checkNicknameUseCase: any CheckNicknameUseCase
     
     init(
         presenter: (any DogInfoInteractorOutput)? = nil,
         saveUserInfoUseCase: any SaveUserInfoUseCase,
         saveProfileImageUseCase: any SaveProfileImageUseCase,
         saveUserInfoRemoteUseCase: any CreateAccountUseCase,
-        signInUseCase: any SignInUseCase
+        signInUseCase: any SignInUseCase,
+        checkNicknameUseCase: any CheckNicknameUseCase
     ) {
         self.presenter = presenter
         self.saveUserInfoUseCase = saveUserInfoUseCase
         self.saveProfileImageUseCase = saveProfileImageUseCase
         self.saveUserInfoRemoteUseCase = saveUserInfoRemoteUseCase
         self.signInUseCase = signInUseCase
+        self.checkNicknameUseCase = checkNicknameUseCase
+    }
+    
+    func signInWithProfileData(dogInfo: UserInfo, imageData: Data?) {
+        Task {
+            do {
+                try await signInUseCase.execute()
+                let fileName = try await saveUserInfoAndProfileImage(dogInfo: dogInfo, imageData: imageData)
+                try await saveUserInfoToRemote(dogInfo: dogInfo, profileImageFileName: fileName)
+                presenter?.didSaveUserInfo()
+            } catch {
+                presenter?.didFailToSaveUserInfo(error: error)
+            }
+        }
     }
     
     private func saveUserInfoAndProfileImage(dogInfo: UserInfo, imageData: Data?) async throws -> String? {
@@ -68,6 +85,7 @@ final class ProfileCreateInteractor: ProfileCreateInteractable {
             return savedFileName
         }
     }
+    
     private func saveUserInfoToRemote(dogInfo: UserInfo, profileImageFileName: String?) async throws {
         guard let userID = try? SupabaseSessionManager.shared.userID.get() else {
             throw SupabaseSessionError.sessionNotExist
@@ -85,20 +103,18 @@ final class ProfileCreateInteractor: ProfileCreateInteractable {
         )
         await saveUserInfoRemoteUseCase.execute(info: userInfoDTO)
     }
+    
     func convertImageToJPGData(image: UIImage?) -> Data? {
         guard let image else { return nil }
         return image.jpegData(compressionQuality: 0.8)
     }
-    func signInWithProfileData(dogInfo: UserInfo, imageData: Data?) {
+    
+    func isNicknameTaken(_ nickname: String) {
         Task {
-            do {
-                try await signInUseCase.execute()
-                let fileName = try await saveUserInfoAndProfileImage(dogInfo: dogInfo, imageData: imageData)
-                try await saveUserInfoToRemote(dogInfo: dogInfo, profileImageFileName: fileName)
-                presenter?.didSaveUserInfo()
-            } catch {
-                presenter?.didFailToSaveUserInfo(error: error)
-            }
+            let isDuplicated = try await checkNicknameUseCase.execute(
+                nickname: nickname
+            )
+            presenter?.notifyNicknameDuplication(isDuplicated)
         }
     }
 }
