+++ SniffMeet/SniffMeet/Source/Scene/Auth/SetProfileInfo/View/ProfileCreateViewController.swift
@@ -1,0 +1,244 @@
+//
+//  ProfileSetupView.swift
+//  SniffMeet
+//
+//  Created by 윤지성 on 11/9/24.
+//
+
+import Combine
+import PhotosUI
+import UIKit
+
+protocol ProfileSetViewable: AnyObject {
+    var presenter: (any ProfileSetPresentable)? { get set }
+    func didFailToCreateProfile()
+    func didSuccessCreateProfile()
+}
+
+final class ProfileSetViewController: BaseViewController, ProfileSetViewable {
+    var presenter: (any ProfileSetPresentable)?
+    private var cancellables = Set<AnyCancellable>()
+    
+    private var titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = Context.mainTitle
+        label.textColor = SNMColor.mainNavy
+        label.numberOfLines = 2
+        label.font = UIFont.systemFont(ofSize: 24, weight: .heavy)
+        return label
+    }()
+    private var profileImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.image = UIImage(named: "ImagePlaceholder")
+        imageView.contentMode = .scaleAspectFill
+        return imageView
+    }()
+    private var addPhotoButton: UIButton = {
+        let button = UIButton()
+        button.setImage(UIImage(systemName: "photo"), for: .normal)
+        button.backgroundColor = SNMColor.mainNavy
+        button.setTitleColor(SNMColor.white, for: .normal)
+        button.tintColor = SNMColor.white
+        return button
+    }()
+    private var warningLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = SNMColor.warningRed
+        label.text = Context.lengthWarning
+        label.font = UIFont.systemFont(ofSize: 12)
+        label.isHidden = true
+        return label
+    }()
+    private var nicknameTextField: InputTextField = InputTextField(placeholder: Context.placeholder)
+    private var submitButton = PrimaryButton(title: Context.submitBtnTitle)
+    private var picker: PHPickerViewController = {
+        var configuration = PHPickerConfiguration()
+        configuration.selectionLimit = 1
+        configuration.filter = .images
+        return PHPickerViewController(configuration: configuration)
+    }()
+    private let snmProgressToast: SNMProgressView = SNMProgressView(
+        animationType: SNMToastAnimation.showAtCenter
+    )
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setSubviewsLayout()
+        view.backgroundColor = SNMColor.white
+        submitButton.isEnabled = false
+        setDelegate()
+        hideKeyboardWhenTappedAround()
+        navigationController?.navigationBar.configureBackButton()
+    }
+    override func viewDidLayoutSubviews() {
+        profileImageView.makeViewCircular()
+        addPhotoButton.makeViewCircular()
+    }
+    override func bind() {
+        addPhotoButton.publisher(event: .touchUpInside)
+            .sink { [weak self] _ in
+                guard let picker = self?.picker else { return }
+                self?.present(picker, animated: true, completion: nil)
+            }
+            .store(in: &cancellables)
+        submitButton.publisher(event: .touchUpInside)
+            .debounce(for: .seconds(EventConstant.debounceInterval), scheduler: RunLoop.main)
+            .sink { [weak self] _ in
+                guard let nickname = self?.nicknameTextField.text else { return }
+                self?.presenter?.didTapSubmitButton(
+                    nickname: nickname,
+                    image: self?.profileImageView.image
+                )
+                self?.submitButton.isEnabled = false
+                self?.snmProgressToast.show(in: self?.view, isDim: true)
+            }
+            .store(in: &cancellables)
+        presenter?.output.isDuplicated
+            .receive(on: RunLoop.main)
+            .sink { [weak self] isDuplicated in
+                guard let text = self?.nicknameTextField.text else { return }
+                let isValidLength = Constants.nameLengthConstraints.contains(text.count)
+                self?.setWarningLabelMessage(
+                    isValidLength: isValidLength,
+                    isDuplicated: isDuplicated
+                )
+                self?.submitButton.isEnabled = !isDuplicated && isValidLength
+                
+            }
+            .store(in: &cancellables)
+    }
+    
+    private func setWarningLabelMessage(isValidLength: Bool, isDuplicated: Bool) {
+        if !isValidLength {
+            warningLabel.text = Context.lengthWarning
+            warningLabel.isHidden = false
+        } else if isDuplicated {
+            warningLabel.text = Context.duplicationWarning
+            warningLabel.isHidden = false
+        } else {
+            warningLabel.isHidden = true
+        }
+    }
+
+    func didSuccessCreateProfile() {
+        Task { @MainActor [weak self] in
+            self?.snmProgressToast.hidden(duration: 0)
+        }
+    }
+    func didFailToCreateProfile() {
+        Task { @MainActor [weak self] in
+            self?.snmProgressToast.hidden { _ in
+                self?.submitButton.isEnabled = true
+            }
+        }
+    }
+}
+
+private extension ProfileSetViewController {
+    func setSubviewsLayout() {
+        [titleLabel,
+         profileImageView,
+         addPhotoButton,
+         warningLabel,
+         nicknameTextField,
+         submitButton].forEach {
+            $0.translatesAutoresizingMaskIntoConstraints = false
+            view.addSubview($0)
+        }
+        
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 16),
+            titleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor,
+                                                constant: Layout.horizontalPadding),
+            profileImageView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            profileImageView.widthAnchor.constraint(equalToConstant: Layout.imageViewSize),
+            profileImageView.heightAnchor.constraint(equalToConstant: Layout.imageViewSize),
+            profileImageView.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 80),
+            addPhotoButton.widthAnchor.constraint(equalToConstant: Layout.addPhotoButtonSize),
+            addPhotoButton.heightAnchor.constraint(equalToConstant: Layout.addPhotoButtonSize),
+            addPhotoButton.trailingAnchor.constraint(equalTo: profileImageView.trailingAnchor,
+                                                     constant: 10),
+            addPhotoButton.bottomAnchor.constraint(equalTo: profileImageView.bottomAnchor),
+            nicknameTextField.topAnchor.constraint(equalTo: profileImageView.bottomAnchor,
+                                                   constant: 60),
+            nicknameTextField.leadingAnchor.constraint(equalTo: view.leadingAnchor,
+                                                       constant: Layout.horizontalPadding),
+            nicknameTextField.trailingAnchor.constraint(equalTo: view.trailingAnchor,
+                                                        constant: -Layout.horizontalPadding),
+            warningLabel.topAnchor.constraint(equalTo: nicknameTextField.bottomAnchor,
+                                              constant: 16),
+            warningLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor,
+                                                  constant: 26),
+            submitButton.leadingAnchor.constraint(equalTo: view.leadingAnchor,
+                                                  constant: Layout.horizontalPadding),
+            submitButton.trailingAnchor.constraint(equalTo: view.trailingAnchor,
+                                                   constant: -Layout.horizontalPadding),
+            submitButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -32)
+        ])
+    }
+    func setDelegate() {
+        picker.delegate = self
+        nicknameTextField.delegate = self
+    }
+}
+
+extension ProfileSetViewController: PHPickerViewControllerDelegate {
+    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
+        picker.dismiss(animated: true, completion: nil)
+        let itemProvider = results.first?.itemProvider
+        if let itemProvider = itemProvider,
+           itemProvider.canLoadObject(ofClass: UIImage.self) {
+            itemProvider.loadObject(ofClass: UIImage.self) { image, error in
+                guard let selectedImage = image as? UIImage else { return }
+                Task { @MainActor [weak self] in
+                    self?.profileImageView.image =  selectedImage
+                }
+            }
+        }
+    }
+}
+
+// MARK: - ProfileCreateViewController+UITextFieldDelegate
+
+extension ProfileSetViewController: UITextFieldDelegate {
+    func textFieldDidBeginEditing(_ textField: UITextField) {
+        submitButton.isEnabled = false
+    }
+    
+    func textFieldDidEndEditing(_ textField: UITextField) {
+        guard let text = textField.text else { return }
+        presenter?.didtextFieldEndEditing(text: text)
+    }
+    
+    func textField(
+        _ textField: UITextField,
+        shouldChangeCharactersIn range: NSRange,
+        replacementString string: String
+    ) -> Bool {
+        guard let text = textField.text else { return false }
+        let newLength = text.count + string.count - range.length
+        let inputTextValid = newLength <= Constants.maxNameLength
+        return inputTextValid
+    }
+}
+
+// MARK: - ProfileCreateViewController+Constants
+
+extension ProfileSetViewController {
+    enum Constants {
+        static let nameLengthConstraints = 2...8
+        static let maxNameLength = 15
+    }
+    enum Context {
+        static let submitBtnTitle: String = "등록 완료"
+        static let placeholder: String = "닉네임을 입력해주세요."
+        static let lengthWarning: String = "닉네임은 최소 2글자, 최대 8글자까지 입력할 수 있습니다."
+        static let duplicationWarning: String = "중복된 닉네임입니다."
+        static let mainTitle: String = "마지막으로,\n반려견 사진과 닉네임을 등록해주세요."
+    }
+    enum Layout {
+        static let horizontalPadding: CGFloat = 24
+        static let imageViewSize: CGFloat = 140
+        static let addPhotoButtonSize: CGFloat = 44
+    }
+}
